% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bam.R
\name{bb_count_bam}
\alias{bb_count_bam}
\title{Count Reads in BAM File}
\usage{
bb_count_bam(path)
}
\arguments{
\item{path}{Character. Path to a BAM file.}
}
\value{
An integer: total number of reads.
}
\description{
Returns the total number of reads in a BAM file using Rsamtools or
system samtools.
}
\examples{
\donttest{
# count <- bb_count_bam("aligned.bam")
}

}
