% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{bb_qc_summary}
\alias{bb_qc_summary}
\title{QC Summary}
\usage{
bb_qc_summary(qc_object)
}
\arguments{
\item{qc_object}{A \code{bb_qc} object from \code{\link[=bb_qc]{bb_qc()}}.}
}
\value{
A data.frame with one row per file.
}
\description{
Returns a summary data.frame of QC metrics.
}
\examples{
\donttest{
tmp <- tempfile(fileext = ".fastq")
writeLines(c("@r1", "ACGT", "+", "IIII"), tmp)
qc <- bb_qc(fastq_path = tmp)
bb_qc_summary(qc)
}

}
