% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{print.bb_qc}
\alias{print.bb_qc}
\title{Print method for bb_qc}
\usage{
\method{print}{bb_qc}(x, ...)
}
\arguments{
\item{x}{A bb_qc object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for bb_qc
}
