% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.bb_result}
\alias{print.bb_result}
\title{Print method for bb_result}
\usage{
\method{print}{bb_result}(x, ...)
}
\arguments{
\item{x}{A bb_result object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for bb_result
}
