% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{summary.bb_qc}
\alias{summary.bb_qc}
\title{Summary method for bb_qc}
\usage{
\method{summary}{bb_qc}(object, ...)
}
\arguments{
\item{object}{A bb_qc object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for bb_qc
}
